Fix wrong assumption about focus in/out event delivery
In 2947d79d assumption was made that we never get focus in event
for popup reason. Unfortunately this is not true as simply
creating window menu and switching between webengineview will trigger
the issue. Moreover, focus in and out events with popup reason do not
mean change of focus, as in qtbase:

 if (QWidget *fw = active_window->focusWidget()) {
     if (fw != QApplication::focusWidget()) {
         fw->setFocus(Qt::PopupFocusReason)
     } else {
         QFocusEvent e(QEvent::FocusIn, Qt::PopupFocusReason);
	 QCoreApplication::sendEvent(fw, &e);
     }
 }

Therefore it is safe to ignore focus in/out events with popup reason.

Note that the fact that ASSERT got not triggered for context menu is
due to another race condition issue in qquick event stack delivery.

This commit amends 2947d79d8210a7ae4ce2bc02e058628e66011be3.

Fixes: QTBUG-101706
Pick-to: dev 6.2
Change-Id: I3e761fe5d8054aed72461eba7981c69755877d1b
Reviewed-by: Kirill Burtsev <kirill.burtsev@qt.io>