Add PdfLinkDelegate instead of link decoration style properties
This simplifies PdfStyle somewhat. PdfLinkDelegate could still be
replaced or augmented via file selectors, but we don't expect to
need to, since link delegates are already non-visual.

[ChangeLog][QtPDF] A PdfLinkDelegate will now be instantiated on top of
each hyperlink in the PdfMultiPageView, PdfScrollablePageView and
PdfPageView components, for event handling and to provide tapped() and
contextMenuRequested() signals. It is non-visual by default, but can be
customized, for example to draw underlines under hyperlinks if the PDF
documents are not expected to have them already.

Change-Id: Ib3726e7220a0c3858db564ca8d53a04625a11950
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>