Support cross-compilation on macOS
So far we only supported a cross-compiling with universal builds,
however qmake also supports 'regular' cross compilation (by setting
'-device-option'). This is semi-supported with qt-cmake as we need to
provide additional configure defines like:

  * CMAKE_OSX_ARCHITECTURES=arm64
  * CMAKE_SYSTEM_NAME=Darwin
  * CMAKE_OSX_DEPLOYMENT_TARGET=10.14

It might seem to be a far fetched issue, however in case of webengine
doing the cross compilation for only one architecture allows to save
compile times when testing only arm64 builds.

Note we do not need to create gn toolchains for that case, however
unlike universal builds it requires host qt build for tools.

Pick-to: 6.3 6.2
Change-Id: Ica8470fdd4cad4866c1470e0403ffd019eaf39a6
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>