PDF multipage example: add footer button to hide/show the sidebar
When the sidebar only held search results, it was enough to show it when
pressing enter on the search field; but now it holds two other views
that the user will want to easily look at. It's a Drawer, and those are
tricky to open and close by mouse-dragging. The button is more obvious,
like the one at the bottom of Creator. Its checked state shows whether
the sidebar is open, and clicking it opens the sidebar if it wasn't
already open. Clicking outside the sidebar closes it.

Change-Id: Ib4e403aa6f06d5e4dbad08a80c99496ab6c5fa0f
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>