Remove QPdfBookmarkModel::structureMode
Bookmarks are naturally a tree. It's unusual for a QAIM to have the
internal ability to flatten a tree to a list, and we don't need it in
any of our viewers or examples. This patch could be reverted if it turns
out that anybody misses this feature; or else flattening could be done
in a proxy model.

Change-Id: I5cf60a39cb699932cc7c61e33a5129323d648344
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>