Add QPdfPageSelector
This is a QSpinBox subclass that shows PDF page "labels" rather than a
1-based or 0-based page index. If a book starts with Roman numerals in
the preface, and then page 1 is the first page of Chapter 1, the spinbox
should be in sync with the page numbers as printed on the pages, and
with the labels that will eventually be shown under the thumbnails
on the Pages sidebar tab.

On the other hand, the user probably needs to see the 1-based page index
somewhere, at least to be able to make sense of the print dialog,
because there the range of pages to print will be 1-based. So we put
the page index into the title bar: title, page label, index, count.

Task-number: QTBUG-102271
Change-Id: Ic461094ba4caae3067409f7f436bd4e7504a4bdb
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>