Fix flakiness of jsServiceWorker test
This is an interesting issue as in the test sometimes first party
url of intercepted request has "sw.js" instead of "sw.html". After
deep debug of the issue the reason behind is that a service
worker registration job checks the registration for given scope
and the key, this in the turn will check storage in
ServiceWorkerStorage::FindRegistrationForScope. Now if registration
is found this will trigger service worker update checker,
which fires the request check for newer version for sw.js and makes
an intercepted request with sw.js as first party url. This of course
brakes the test.

Do not use the storage in that test, and let the service worker
registration job do full "install"  & "register" cycle instead
of running "update" phase.

As a side note , using off the record profile makes test run faster
as opening the storage and closing it is costly.

Pick-to: 6.3 6.2
Change-Id: If13a4fcf0014cc5c76cb8ed14e60f2e5e29c67ef
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>