Add link role to QPdfLinkModel, providing a QPdfLink instance
We want to call PdfPageNavigator.jump(link) when a link is clicked;
therefore PdfLinkModel needs to provide this object; therefore it makes
sense to store the links as QPdfLink instances, and get rid of the
redundant internal Link object type. But QPdfLink did not have storage
for a QUrl until now. The textStart and textCharCount fields sounded
useful, but in practice were set and not used.

QPdfLink gains toString() and copyToClipboard() methods.

Also removed some unused cruft in qquickpdflinkmodel.

[ChangeLog][QtPDF] PdfLinkModel now provides a QPdfLink object for each
link. QPdfLink now contains everything necessary to render delegates for
links and search results, and handle clicking links; and there is a
copyToClipboard() method for use in context menus, which will copy
the text returned trom toString(), which is also invokable.

Change-Id: Ifa5abe6df8f8d6f9ef98971044c26691ec4b9a36
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>