PdfLinkModel: if quadpoints available, convert to rects
If QPdfLink stored polygons instead of rects, it would make sense to
transform each point using QPdfDocumentPrivate::mapPageToView();
but the quadpoints (polygons) in the test.pdf from QTBUG-100630 happen
to describe rectangles. This patch successfully does the transform so
that they line up with the link text. It doesn't do much good since the
links in that PDF are invalid anyway, but presumably links using
quadpoints occur in the real world sometimes.

Pick-to: 6.7
Task-number: QTBUG-100630
Change-Id: I0cb151dbb30ca73c8f9ed2bcd29508ced05de064
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>