Fix crashes of combobox popup tests
Offscreenwindow from quickwidget is a top level window,
which was from time to time used as "popup" window in the test
instead of widgetwindow of webengineview.
Check if window is offscreenwindow and do not consider it
as valid window in the test.
Moreover popup should be closed in the end of test, however
checking for offscreen window would result in bogus outcome
as offscreen window is destroyed before quickwidgetwindow.

Task-number: QTBUG-105342
Change-Id: Ibfeab0cad0bc43690c2458cb97738b4b5ea132f4
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>
(cherry picked from commit 7415a127df0ab21b703d1a549abe595900e76fab)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>