Build the QtDesigner plugin in all configurations
The QtDesigner plugin was marked as 'tool_plugin', which restricts its
build to release-only, even in debug_and_release builds.  This was done
to avoid building debug plugins for tools that are only build as
release.  However, the designer plugins are also loaded when using
QUiLoader and thus are not exclusively meant for the designer tool.

Task-number: QTBUG-104755
Change-Id: Id87012f57ad06984c7b0f0d318f75ad2c9a596f2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Michal Klocek <michal.klocek@qt.io>