Native spellchecker: Fix it when disabled
SpellCheck is enabled by default "due to peculiarities of the
initialization sequence" (see the ctor in
components/spellcheck/renderer/spellcheck.cc). It produces red underlines
on the fist user interaction even if we disable it from WebEngineProfile,
then the lazily initialized SpellcheckService disables it correctly.

Initialize the service early for this particular case to avoid undesired
spellchecking.

Change-Id: Ib2e38fb381d9b169f8f59c8100b4df7af3493aa1
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>