Add workaround for un-minimizing QWebEngineView under Gnome
Gnome WM doesn't send expose event when minimizing then restoring a
window. Presumably, due to the missing expose event the top-level
QWebEngineView is not redrawn. As a workaround, force to update
RenderWidgetHostViewQtDelegateWidget when shown.

The workaround is not needed in 6.4 and later because the issue is fixed
by 5d1ef38f9 Create a RWHV delegate in core

Pick-to: 5.15
Fixes: QTBUG-104763
Change-Id: I4761d670d17f6dbbe3e0de82a00179eb7fd7913c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>