Reduce slowness and flakiness of tst_origin
This test runs very slow and it is flaky. Make it run faster by:

* instead of doing timeout checks for "cannotLoad", introduce
<image> element add use 'onload' and 'onerror' handlers before
doing actual <iframe> load, this dropps a need for hardcoded
delay/deadline for ifrme being loaded.
* do not evaluate JavaScript in try wait loop , use JavaScript
messages instead.

Note that using <img> or <script> element to perform the test (which
have onload and onerros handler unlike iframe) ie. loading cross origin
scripts or images instead of iframes is a dead-end as it has different
results and "canLoadButNotAccess" can not be tested as origin BAR can
load a script from origin FOO, but it still works in BAR‘s context.

Task-number: QTBUG-105342
Pick-to: 6.4
Change-Id: I180aea801cedb0f97b96cb90a54b793d0a7c1a7f
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>