CMake: accept known valid 32 bit host architectures
When trying to cross-build from one 32 bit host architecture to another the
"Unsupported architecture:" case was hit, in which case the result of
get_v8_arch() was empty. This would lead to configure_gn_toolchain() being
called with a missing argument, which gave a totally unhelpful error message.

Fix this by just accepting valid 32 bit architectures from the internal list,
and raise the debug message to a FATAL_ERROR to get immediate feedback.

Pick-to: 6.4 6.3
Change-Id: I8e0c757ad13e2acff67827bd9063c493e9d84444
Reviewed-by: Michal Klocek <michal.klocek@qt.io>