PdfSelection: mark read-only for IM to avoid showing virtual keyboard
The Qt::ImReadOnly query enum is new in 6.2 (qtbase
c80f262258b7846bf199887bcfdbb6dcfda6ad6f and
aae4d52cbb79c61161271473ab5c3a1ba1086e01). It was added to the
switch statement in 79d04aa9e3e9aa84d7378260519f9e9a6759dc41 but
needs to return `true` to mark this Item as read-only, so that the
keyboard will not open when it gets focus. The menu still gets
populated with Select and Select All on long-press, and then
Copy after a selection is made. That's all we need.

The qt_im_readonly property was a sort of private API for Qt 5, and
should no longer be needed (albeit the implementation is still in Qt 6
so far).

Pick-to: 6.4
Task-number: QTBUG-83811
Fixes: QTBUG-106358
Change-Id: I677363be545d07884dffdfc10d6fdbd488cf2cf0
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>