Do not let print engine multiply number of copies
We print multiple pages by repeatedly drawing in a loop, which works good
for PDF targets, but fails for physical printing. PPK_CopyCount flag is
provided to the printing backend which multiplies our effort in this
case.

Temporarily setting the print engine back to 1 seems to be more
reliable than trying to distinguish between paper of PDF target.

Task-number: QTBUG-104224
Change-Id: Ic1be15c6965e6a739bcd44c8d904ac19f67719b5
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>