Return both application and system certificates
For the certificate choice return both application and
system certificates.

Add unit test to cover the case on Linux.
Unfortunately it requires adding the user certificate to
the nss data store, which is not not nice, however porting
the certificate manger from Chromium is a bigger task.

Test runs only if the machine has pk12utils installed.

During the test the user certificate is imported into
the nss database with the nickname 'qwebengineclientcertificatestore'.

This can be removed later with:
ninja remove-user-personal-certificate

and verified with:
certutil -d sql:$HOME/.pki/nssdb -L

Pick-to: 6.4
Change-Id: I475fddc68ea56304980f6c835ed4cfed4b093ad4
Reviewed-by: Michael Brüning <michael.bruning@qt.io>