Fix bogus layouting of qwebengineview
We did not really change pointers for doing a widget
and a item delegate binding. This at some point was
blowing up as we did not remove the old widget before
inserting the new widget into the internal webengineview
layout. The issue could get unnoticed for a while as the
old widget got eventually removed on a destruction.
However, at times when on the qlayout activation call
the old widget was considered, resize event
was triggered which was setting viewport to false value
(minimum size to hold two webcontents).

Make sure we remove widget before adding the new widget.
This fixes the widget and the item bind, however this
logic should be rewritten for dev branch as is get more
and more complex.

Fixes: QTBUG-107529
Pick-to: 6.4
Change-Id: I31d180b54f52e38b3c335cfa2bf3afa61e449677
Reviewed-by: Martin Negyokru <negyokru@inf.u-szeged.hu>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>