Mention WebSockets in docs
QtWebChannel has an example for a use case of HTML web sockets and
connectivity of the native side of application. We don't need to create
a new example of our own by mentioning this.

However, the chat client application uses JS libraries served from
remote, so we have to enable LocalContentCanAccessRemoteUrls setting
to make it work out of the box in WebEngine example browsers. Desktop
Chrome works the same way by default.

Task-number: QTBUG-106505
Change-Id: I9020d0a09a88de16d32af063aee5d55c9837f484
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>