Deprecate Quota Permission Request API
Requesting host quota is no longer supported by Chromium.
navigator.webkitPersistentStorage has became an alias for
navigator.webkitTemporaryStorage after the Chromium 106 update.
Requesting quota for temporary storage is needless because the
allocation is automatic and storage can't be requested beyond the
maximum limit.

The logic of the persistent storage will be entirely removed in the next
Chromium update so just deprecate the API and remove the corresponding
implementations.

[ChangeLog][QtWebEngineCore] Deprecate QWebEnginePage::quotaRequested()
signal and QWebEngineQuotaRequest class. The signal is not emmitted
anymore.
[ChangeLog][QtWebEngineQuick] Deprecate QWebEngineView.quoataRequested()
signal. The signal is not emitted anymore.

Task-number: QTBUG-56354
Change-Id: Ie397598c416d9e811270ba56a97f8984948277f7
Reviewed-by: Szabolcs David <davidsz@inf.u-szeged.hu>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>