Disable V8 sandbox for QEMU arm64 CI
After the 106-based adaptations V8 sandbox is enabled by default. It
cannot be disabled runtime anymore.

The initialization of the sandbox has significant overhead in a
qemu-aarch64 environment, at least 5sec. This makes most of our auto
tests timeout.

The sandbox is not supported on 32bit platforms. The detection of the
QEMU target relies on the environment variable
TARGET_OSVERSION_COIN=qemu on the CI.

Fixes: QTBUG-108993
Change-Id: I45961bc2e8a0af505113777f2b26f10b93b7477f
Reviewed-by: Michal Klocek <michal.klocek@qt.io>