Fix opening files with File System Access API
Chrome interprets accepting file picker dialog as a user consent and
grants access permission to the selected files automatically. We have
to do the same since we don't get any permission request in this case
anymore.

This issue did not affect directory picker and file saver modes.

Also add AncestorHasActivePermission() check, it seems to be effective
to not request permission again for the same file if the parent dir
was already pulled in by this API.

Change-Id: Id73d8fc6e9bd518692362133f3dafa472f0e97a1
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 01728421d641ffac98a85b408167db478b253458)
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>