Avoid scaling mouse movement properties on High DPI screens
ui::TranslateAndScaleWebInputEvent() scales movementX/Y if the input
event is not raw.

It is not exactly clear what is the expected behavior and browsers
handle movementX/Y scaling differently. W3C states that
currentEvent.movementX = currentEvent.screenX - previousEvent.screenX
which is not true with scaling.

This is a workaround for https://crbug.com/907309

This fixes tst_QWebEnginePage::mouseMovementProperties auto test with
QT_SCALE_FACTOR=2.

Pick-to: 6.5
Change-Id: I6d1fd591ec2bbeb22c448772c367b59509244491
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>