Skip favicon database QML tests on High DPI screens
The tests use the middle pixel's value of an image to check if a
certain icon is loaded. The images are scaled on High DPI screens and the
pixel values are changed due to interpolation. It cannot be expected to
have the same pixel values for different device-pixel-ratio settings,
thus rather skip the test than maintaining a list of pixel values for
different scales.

Pick-to: 6.5
Change-Id: I8d142de0106167a5d16c540fa468003b36c95317
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>