Skip favicon database QML tests on High DPI screens
The tests use the middle pixel's value of an image to check if a
certain icon is loaded. The images are scaled on High DPI screens and the
pixel values are changed due to interpolation. It cannot be expected to
have the same pixel values for different device-pixel-ratio settings,
thus rather skip the test than maintaining a list of pixel values for
different scales.

Change-Id: I8d142de0106167a5d16c540fa468003b36c95317
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 86da28c5658230a361cd94a514d500c3ad3a7de9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>