RenderWidgetHostViewQtDelegateItem: keep grabs
If a WebView or WebEngineView gets a mouse or touch grab, we don't want
to give it up to any handlers that may be lurking on Items underneath.
WebView is opaque, so we don't know whether the user is trying to drag a
scrollbar or something else, but interactive items outside shouldn't
interfere, in general.

[ChangeLog][QtWebEngineQuick] WebEngineView (or WebView backed by Qt
WebEngine) no longer allows components outside to take over the mouse or
touch exclusive grab. For example if the user starts dragging a
scrollbar inside the web view, that continues until release, regardless
of any DragHandler, Flickable etc.

Fixes: QTBUG-103518
Change-Id: I4352dc8482020f0efc7a0901e94ccf7fc147fa1b
Reviewed-by: Michal Klocek <michal.klocek@qt.io>
(cherry picked from commit 3feab861f5ead63e2a661343f0df2358fa0af894)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>