Improve error handling of argument parsing
As a corner case, QCoreApplication::arguments() might be empty. For
example, the embedder sets argc=0. It is invalid but doesn't crash
or warn.

base::CommandLine expects program name to be set and Chromium code might
use it. It is not possible to set program name if argv is not passed to
QCoreApplication.

This change does not handle this corner case but detects it, and warns
the user to not expect proper behavior.

Pick-to: 6.5
Task-number: QTBUG-110157
Change-Id: Ibf14b11bbf8b8c72d8a1d8419377a25b311b9ebe
Reviewed-by: Michal Klocek <michal.klocek@qt.io>