QWebEngineLoadingInfo: Add response headers
This change adds a member variable that contains the response headers
to QWebEngineLoadingInfo. It is filled when a
QWebEngineLoadingInfo instance is constructed inside
WebContentsDelegateQt::emitLoadFinished(bool).

The response headers are extracted from the navigation handle when
WebContentsDelegateQt::DidFinishNavigation(content::NavigationHandle *)
is called.

The response headers are non-empty when QWebEngineLoadingInfo::status()
is equal to QWebEngineLoadingInfo::LoadSucceededStatus or
QWebEngineLoadingInfo::LoadFailedStatus.

Fixes: QTBUG-106862
Change-Id: I4d196e3cc71725ddad9a5832af72d1b4e50924c8
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Michael Brüning <michael.bruning@qt.io>