Add Origin Trial Policy support
This makes SharedArrayBuffer available for earth.google.com

For enabling certain experimental feature for a web page, the
page provides a token in a meta tag for the browser. The token
is requested by the web developer and generated by Google. The
verification of the token happens without any server calls or
network access. See,
 * http://googlechrome.github.io/OriginTrials/developer-guide.html#23-how-are-tokens-verified
 * blink::TrialToken::Parse() and blink::TrialTokenValidator
   implementation (//third_party/blink/common/origin_trials)

Pick-to: 6.5
Change-Id: I6f102af8efcd12b6b267dceb9de59d685fd22f9e
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>