Add cipd handling to init-repository and take_snapshot
Add parsing deps for packages deployed by cipd.
Keep changes to minimum as we so far need just one
androidx package. All the package files are filtered by the
chromium black list.

Whitelist pydeps and info files for android dependencies,
as this minimizes required changes to gn build files,
add icu file for android.

Update the version to 108.0.5359.181 as it was outdated.

Cipd package parsing is done only for androidx package and
the logic runs for

* init-repository --baseline-upstream
  > cipd downloads and extracts androidx package to
    upstream repository
* take-snapshot
  > cipd returns content so it can got through blacklist lists
    (files are not in git so git ls-file does not help here)

Task-number: QTBUG-83459
Change-Id: I7b8203bfb273ba90113a449725ba9ea5c9a6692c
Reviewed-by: Michael Brüning <michael.bruning@qt.io>