Fix uniqueId() for accessible events
QAccessibleEvent::uniqueId() expects child index to be set if the event
object is set. If the index is not explicitly set the event interprets the
unique ID as a child index what is invalid.

This change fixes warnings like:
 qt.accessibility.core: Invalid child in QAccessibleEvent:
 QObject(0x1cc9ec25d60, name = "search") child: -2147483017

In case of an invalid child, uniqueId() returns 0 and this causes assert
on macOS.

Task-number: QTBUG-105988
Change-Id: If068ab648226284bb3bd6502fc0381f4229d19e2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 442e347f592d389ff664f3b1cae466c3e2506d78)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>