Teach QPdfView to render search result highlights
The bounding boxes of all visible search results are filled with a
translucent color on top of the page image, because we don't know
whether the image is opaque or not. currentSearchResult is further
highlighted by drawing bounding box borders.

The colors are the same as PdfStyle.pageSearchResultsColor and
PdfStyle.currentSearchResultStrokeColor, respectively. Perhaps they
could be configurable somehow in the future, but it's not clear yet what
API that ought to have, since we do not have suitable QPalette roles
(except QPalette::Highlight, which is not translucent).

Task-number: QTBUG-77507
Task-number: QTBUG-92973
Change-Id: I3df634229811e72d2d438a39997e0b4887d6bf11
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>