Add Windows 11 related note for Geolocation settings
There are 3 OS settings under Windows 11 that need to be enabled for
the HTML5 Geolocation in QtWebEngine to work properly. This isn't
very straightforward to a user, therefore this patch adds a note to
the documentation.

Change-Id: Id36a4726a182178c5a15fce039e464fd6bb7517e
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit c3b2393e7a532a18d4971a878740b2a9b5830028)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>