Enable accessibility by default on Linux
Previously this was disabled unless the
QTWEBENGINE_ENABLE_LINUX_ACCESSIBILITY was set, as the debian packaging
was always enabling accessibility[1] even if a screen reader was not
enabled.

This is not the case anymore since 5 years ago[2] and now accessibility will
only be enabled if a screen reader is detected. Which is the correct
upstream behavior of at-spi2-core.

So now enable accessibility unless
QTWEBENGINE_ENABLE_LINUX_ACCESSIBILITY is set to 0

[1]: https://salsa.debian.org/a11y-team/at-spi2-core/-/commit/2a99b7b40526bbdf091cc574ec08d86c9a46f405
[2]: https://salsa.debian.org/a11y-team/at-spi2-core/-/commit/fde0bbead6aacefd0b5dcf9f6d36f7f50a2c0f3d

Pick-to: 6.5 5.15
Change-Id: Ie5554f8b578dcca87a54dad525ae6a83d6f8f9d8
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>