Use paintRect instead of pageRect when printing to account for margins
Currently, when printing via QWebEngineView::print(), the margins are handled by QPainter and not Chromium, unlike QWebEngineView::printToPdf() or QWebEnginePage::printToPdf(). The changes in this patch don't affect the
latter two.

QPainter handles margins by making a distinction between pageRect and
paintRect, where paintRect == pageRect - margins. Furthermore, to
actually apply the top and left margin, QPainter moves paintRect by
(marginLeft, marginTop).

This means that we have to draw what is to be printed in the size of
paintRect and QPainter will move it to apply top and left margins.
Right now, this is not the case, as we draw what is to be printed in
the size of pageRect. QPainter then moves it, thereby the result
contains the top and left margins, but overflows the page on the
right and bottom, effectively cropping the final result.

In this patch we fix this by using the size of paintRect instead of
pageRect to draw what is to be printed when printing via
QWebEngineView::print().

Fixes: QTBUG-111909
Change-Id: Ibc7a9035e654e92d84a594c6dd1ec2efc6e108ff
Reviewed-by: Michal Klocek <michal.klocek@qt.io>
(cherry picked from commit 421d3c4e0b57170343df57de0b222d0f57a7bcb7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>