Fix download button of PDF viewer plugin
The basic case of saving a PDF is easily fixed by installing
PDFWebContentsHelpers in ExtensionsAPIClientQt for the proper
WebContents.

If the PDF file contains input elements and the user modifies
them, the download button offers an option to save the modified
file instead of the original. We need a minimal implementation
of FileSystemDelegate, because this happens through file system
access dialogs instead of normal "save as" dialogs.

Pick-to: 6.6
Task-number: QTBUG-104610
Task-number: QTBUG-113149
Change-Id: I756f9fe6c72ccb129b04282b2fda05602cbf54cd
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>