Generate chromium_attributions.qdoc before running qdoc -prepare
If chromium_attributions.qdoc doesn't exist by the time QDoc is run
in -prepare mode, Chromium attribution pages are not listed in
qtwebengine.index file and therefore not visible to documentation sets
outside Qt WebEngine.

This meant that the attributions were not listed on the main Qt
Licensing documentation page (that lives in qtdoc).

Make the prepare_docs_WebEngineCore target depend on
generate_chromium_attributions to ensure the correct order for
doc generation.

Pick-to: 6.5
Fixes: QTBUG-113806
Change-Id: I4e2a16974d1c6d29c3942becfbb46f0078cf3113
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>