Fix crashes on sendEvent on 'invalid' key.text
A user can create QKeyEvent which may have too long
QKeyEvent::text and does not fit into null terminated blink
WebKeyboardEvent::text. Till now the text was mem copied
however without null-termination if it was too long
(which is char16_t[WebKeyboardEvent::kTextLengthCap])

It seems that real keyboard events do not generate any 'key.text'
for modifiers, however user can provide any text (for example
in the form of 'shift' 'control'). Therefore 'alt' still
worked as we had 16 bytes of space but 'shift' was already
too long.

In case we can not fit into blink web keyboard event text
with null terminated string simply skip the text
as it is not there in the first place in case of non-user
generated key events. This way the text is always
null-terminated.

Pick-to: 6.6 6.5 5.15
Fixes: QTBUG-113704
Change-Id: Ib1560a44e7451ab7d886c72338b7433fc51318ed
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>