Fix invalid html in tst_touchinput
With the invalid 'width' attribute the input element
defaults to a size that depends on the system (charmap) settings.
It may result in failing tests that depend on the size of the element
e.g. tests that use 'elementCenter'.

Task-number: QTBUG-105342
Change-Id: I508d365fc4240d67d3e2240e8e33495c1de6ee4c
Reviewed-by: Michal Klocek <michal.klocek@qt.io>
(cherry picked from commit 17bafdc301a057ed4b28dc3acc9bd636183683d0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>