Fix PDF page label search: don't trim in QPdfDocument::pageIndexForLabel()
In the 6.6 header review it was suggested that the caller should do the
trimming if necessary; avoiding it here is more efficient. So far the
only caller is QPdfPageSelector (although we may need to implement the
same feature in the Quick PDF examples at some point).

Typing page labels into QPdfPageSelector didn't work because
QPdfPageSelector::validate() was not overridden. Now we reuse
pageIndexForLabel() for that too: if the label is found, it's valid; if
the label is not found, we just assume the user is not done typing yet.
There's no QPdfDocument function to do a substring search for page
labels, or to return the whole list of them either. It seems they
can only be gotten one-at-a-time via the FPDF_GetPageLabel function.

Now it actually works in the widget-based example
qtwebengine/examples/pdfwidgets/pdfviewer/pdfviewer: if you load a
document that has some Roman-numeral-numbered preface pages and type
e.g. " iii" into the spinbox, it jumps to page iii.

Fixes: QTBUG-115476
Pick-to: 6.6
Change-Id: I070338effee1bbf916581c5ba964367424b786b2
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>